+++ common/src/main/java/com/chefmooon/frightsdelight/common/CommonSetup.java
@@ -24,5 +24,22 @@ public class CommonSetup {
         CompostingChanceRegistry.INSTANCE.add(BuiltInRegistries.ITEM.get(FrightsDelightItems.COOKIE_SLIME).get().value(), 0.85F);
         CompostingChanceRegistry.INSTANCE.add(BuiltInRegistries.ITEM.get(FrightsDelightItems.COOKIE_COBWEB).get().value(), 0.85F);
         CompostingChanceRegistry.INSTANCE.add(BuiltInRegistries.ITEM.get(FrightsDelightItems.COOKIE_GHAST_TEAR).get().value(), 0.85F);
+
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.ROTTEN_FLESH_PIE_SLICE).get().value(), 0.85F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.SLIMEAPPLE_PIE_SLICE).get().value(), 0.85F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.SPIDEREYE_PIE_SLICE).get().value(), 0.85F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.GHASTTEAR_PIE_SLICE).get().value(), 0.85F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.SOUL_BERRY_CHEESECAKE_SLICE).get().value(), 0.85F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.WITHER_BERRY_CHEESECAKE_SLICE).get().value(), 0.85F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.COBWEB_PIE_SLICE).get().value(), 0.85F);
+
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.ROTTEN_FLESH_PIE).get().value(), 1.0F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.SLIMEAPPLE_PIE).get().value(), 1.0F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.SPIDEREYE_PIE).get().value(), 1.0F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.GHASTTEAR_PIE).get().value(), 1.0F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.SOUL_BERRY_CHEESECAKE).get().value(), 1.0F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.WITHER_BERRY_CHEESECAKE).get().value(), 1.0F);
+        ComposterBlock.COMPOSTABLES.put(BuiltInRegistries.ITEM.get(FrightsDelightItems.COBWEB_PIE).get().value(), 1.0F);
+
     }
 }