+++ common/src/main/java/com/chefmooon/frightsdelight/common/tag/CommonTags.java
@@ -36,6 +36,8 @@ public class CommonTags {
     public static final TagKey<Item> C_FOODS_SOUP = getCommonItemTagKey("foods/soup");
     public static final TagKey<Item> C_FOODS_COOKIE = getCommonItemTagKey("foods/cookie");
 
+    public static final TagKey<Item> C_FOODS_MILK = getCommonItemTagKey("foods/milk");
+
     public static final TagKey<Item> C_TOOLS = getCommonItemTagKey("tools");
     public static final TagKey<Item> C_TOOLS_KNIFE = getCommonItemTagKey("tools/knife");
     public static final TagKey<Block> C_MINEABLE_KNIFE = getCommonBlockTagKey("mineable/knife");