+++ fabric/src/main/java/com/chefmooon/frightsdelight/data/fabric/AdvancementGenerator.java
@@ -151,6 +151,13 @@ public class AdvancementGenerator extends FabricAdvancementProvider {
                 .addCriterion("has_punch_ghasttear", InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.PUNCH_GHASTTEAR.get()))
                 .addCriterion("has_punch_soul_berry", InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.PUNCH_SOUL_BERRY.get()))
                 .addCriterion("has_punch_wither_berry", InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.PUNCH_WITHER_BERRY.get()))
+                .addCriterion(RecipeProvider.getHasName(FrightsDelightItemsImpl.ROTTEN_FLESH_PIE_SLICE.get()), InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.ROTTEN_FLESH_PIE_SLICE.get()))
+                .addCriterion(RecipeProvider.getHasName(FrightsDelightItemsImpl.SLIMEAPPLE_PIE_SLICE.get()), InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.SLIMEAPPLE_PIE_SLICE.get()))
+                .addCriterion(RecipeProvider.getHasName(FrightsDelightItemsImpl.SPIDEREYE_PIE_SLICE.get()), InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.SPIDEREYE_PIE_SLICE.get()))
+                .addCriterion(RecipeProvider.getHasName(FrightsDelightItemsImpl.GHASTTEAR_PIE_SLICE.get()), InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.GHASTTEAR_PIE_SLICE.get()))
+                .addCriterion(RecipeProvider.getHasName(FrightsDelightItemsImpl.SOUL_BERRY_CHEESECAKE_SLICE.get()), InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.SOUL_BERRY_CHEESECAKE_SLICE.get()))
+                .addCriterion(RecipeProvider.getHasName(FrightsDelightItemsImpl.WITHER_BERRY_CHEESECAKE_SLICE.get()), InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.WITHER_BERRY_CHEESECAKE_SLICE.get()))
+                .addCriterion(RecipeProvider.getHasName(FrightsDelightItemsImpl.COBWEB_PIE_SLICE.get()), InventoryChangeTrigger.TriggerInstance.hasItems(FrightsDelightItemsImpl.COBWEB_PIE_SLICE.get()))
                 .requirements(AdvancementRequirements.Strategy.AND)
                 .save(consumer, getAdvancementName("frd_master"));
     }