+++ changelog.md
@@ -9,6 +9,22 @@ This mod is compatible with Minecraft 1.21.5 and Farmer's Delight Refabricated
 
 Minecraft 1.21.5 Fabric Release!
 
+### Added
+
+- New Food
+    - Rotten Flesh Pie (+Slice)
+    - Slime Apple Pie (+Slice)
+    - Spider Eye Pie (+Slice)
+    - Ghast Tear Pie (+Slice)
+    - Soul Berry Cheesecake (+Slice)
+    - Wither Berry Cheesecake (+Slice)
+    - Cobweb Pie (+Slice)
+- Origins Tags
+    - Origins Meat: Slice of Rotten Flesh Pie, Slice of Slime-Apple Pie, Slice of Spider Eye Pie
+    - Ignore Diet: Ghast Tear Pie Slice, Cobweb Pie Slice
+- Create Tag
+    - Upright on Belt: Rotten Flesh Pie, Slime-Apple Pie, Spider Eye Pie, Ghast Tear Pie, Soul Berry Cheesecake, Wither Berry Cheesecake, Cobweb Pie
+
 ### Fixed
 
 - Added missing Slime Apple Soup to Master of Fright advancement